requirements.txt
python-telegram-bot==22.6
aiosqlite==0.20.0
openpyxl==3.1.5
python-dotenv==1.0.1
